+++ internal/bundler/linker.go
@@ -46,6 +46,10 @@ type linkerContext struct {
 	// We may need to refer to the CommonJS "module" symbol for exports
 	unboundModuleRef js_ast.Ref
 
+	// We may need to refer to the "__esm" and/or "__commonJS" runtime symbols
+	cjsRuntimeRef js_ast.Ref
+	esmRuntimeRef js_ast.Ref
+
 	// This represents the parallel computation of source map related data.
 	// Calling this will block until the computation is done. The resulting value
 	// is shared between threads and must be treated as immutable.
@@ -227,6 +231,16 @@ func link(
 		),
 	}
 	timer.End("Clone linker graph")
+
+	// Use a smaller version of these functions if we don't need profiler names
+	runtimeRepr := c.graph.Files[runtime.SourceIndex].InputFile.Repr.(*graph.JSRepr)
+	if c.options.ProfilerNames {
+		c.cjsRuntimeRef = runtimeRepr.AST.NamedExports["__commonJS"].Ref
+		c.esmRuntimeRef = runtimeRepr.AST.NamedExports["__esm"].Ref
+	} else {
+		c.cjsRuntimeRef = runtimeRepr.AST.NamedExports["__commonJSMin"].Ref
+		c.esmRuntimeRef = runtimeRepr.AST.NamedExports["__esmMin"].Ref
+	}
 
 	for _, entryPoint := range entryPoints {
 		if repr, ok := c.graph.Files[entryPoint.SourceIndex].InputFile.Repr.(*graph.JSRepr); ok {
@@ -1840,8 +1854,7 @@ func (c *linkerContext) createWrapperForFile(sourceIndex uint32) {
 	// of it.
 	case graph.WrapCJS:
 		runtimeRepr := c.graph.Files[runtime.SourceIndex].InputFile.Repr.(*graph.JSRepr)
-		commonJSRef := runtimeRepr.AST.NamedExports["__commonJS"].Ref
-		commonJSParts := runtimeRepr.TopLevelSymbolToParts(commonJSRef)
+		commonJSParts := runtimeRepr.TopLevelSymbolToParts(c.cjsRuntimeRef)
 
 		// Generate the dummy part
 		dependencies := make([]js_ast.Dependency, len(commonJSParts))
@@ -1863,7 +1876,7 @@ func (c *linkerContext) createWrapperForFile(sourceIndex uint32) {
 			Dependencies: dependencies,
 		})
 		repr.Meta.WrapperPartIndex = ast.MakeIndex32(partIndex)
-		c.graph.GenerateSymbolImportAndUse(sourceIndex, partIndex, commonJSRef, 1, runtime.SourceIndex)
+		c.graph.GenerateSymbolImportAndUse(sourceIndex, partIndex, c.cjsRuntimeRef, 1, runtime.SourceIndex)
 
 	// If this is a lazily-initialized ESM file, we're going to need to
 	// generate a wrapper for the ESM closure. That will end up looking
@@ -1877,8 +1890,7 @@ func (c *linkerContext) createWrapperForFile(sourceIndex uint32) {
 	// for similar reasons to the CommonJS closure above.
 	case graph.WrapESM:
 		runtimeRepr := c.graph.Files[runtime.SourceIndex].InputFile.Repr.(*graph.JSRepr)
-		esmRef := runtimeRepr.AST.NamedExports["__esm"].Ref
-		esmParts := runtimeRepr.TopLevelSymbolToParts(esmRef)
+		esmParts := runtimeRepr.TopLevelSymbolToParts(c.esmRuntimeRef)
 
 		// Generate the dummy part
 		dependencies := make([]js_ast.Dependency, len(esmParts))
@@ -1898,7 +1910,7 @@ func (c *linkerContext) createWrapperForFile(sourceIndex uint32) {
 			Dependencies: dependencies,
 		})
 		repr.Meta.WrapperPartIndex = ast.MakeIndex32(partIndex)
-		c.graph.GenerateSymbolImportAndUse(sourceIndex, partIndex, esmRef, 1, runtime.SourceIndex)
+		c.graph.GenerateSymbolImportAndUse(sourceIndex, partIndex, c.esmRuntimeRef, 1, runtime.SourceIndex)
 	}
 }
 
@@ -2811,8 +2823,8 @@ func (c *linkerContext) findImportedFilesInCSSOrder(entryPoints []uint32) (exter
 	}
 
 	// Include all files reachable from any entry point
-	for _, entryPoint := range entryPoints {
-		visit(entryPoint, ast.Index32{})
+	for i := len(entryPoints) - 1; i >= 0; i-- {
+		visit(entryPoints[i], ast.Index32{})
 	}
 
 	// Reverse the order afterward when traversing in CSS order
@@ -3525,8 +3537,6 @@ func (c *linkerContext) generateCodeForFileInChunkJS(
 	partRange partRange,
 	entryBits helpers.BitSet,
 	chunkAbsDir string,
-	commonJSRef js_ast.Ref,
-	esmRef js_ast.Ref,
 	toModuleRef js_ast.Ref,
 	runtimeRequireRef js_ast.Ref,
 	result *compileResultJS,
@@ -3618,7 +3628,7 @@ func (c *linkerContext) generateCodeForFileInChunkJS(
 				cjsArgs = []js_ast.Expr{{Data: &js_ast.EArrow{Args: args, Body: js_ast.FnBody{Stmts: stmts}}}}
 			}
 			value := js_ast.Expr{Data: &js_ast.ECall{
-				Target: js_ast.Expr{Data: &js_ast.EIdentifier{Ref: commonJSRef}},
+				Target: js_ast.Expr{Data: &js_ast.EIdentifier{Ref: c.cjsRuntimeRef}},
 				Args:   cjsArgs,
 			}}
 
@@ -3686,7 +3696,7 @@ func (c *linkerContext) generateCodeForFileInChunkJS(
 				esmArgs = []js_ast.Expr{{Data: &js_ast.EArrow{Body: js_ast.FnBody{Stmts: stmts}, IsAsync: isAsync}}}
 			}
 			value := js_ast.Expr{Data: &js_ast.ECall{
-				Target: js_ast.Expr{Data: &js_ast.EIdentifier{Ref: esmRef}},
+				Target: js_ast.Expr{Data: &js_ast.EIdentifier{Ref: c.esmRuntimeRef}},
 				Args:   esmArgs,
 			}}
 
@@ -4297,8 +4307,6 @@ func (c *linkerContext) generateChunkJS(chunks []chunkInfo, chunkIndex int, chun
 	chunkRepr := chunk.chunkRepr.(*chunkReprJS)
 	compileResults := make([]compileResultJS, 0, len(chunkRepr.partsInChunkInOrder))
 	runtimeMembers := c.graph.Files[runtime.SourceIndex].InputFile.Repr.(*graph.JSRepr).AST.ModuleScope.Members
-	commonJSRef := js_ast.FollowSymbols(c.graph.Symbols, runtimeMembers["__commonJS"].Ref)
-	esmRef := js_ast.FollowSymbols(c.graph.Symbols, runtimeMembers["__esm"].Ref)
 	toModuleRef := js_ast.FollowSymbols(c.graph.Symbols, runtimeMembers["__toModule"].Ref)
 	runtimeRequireRef := js_ast.FollowSymbols(c.graph.Symbols, runtimeMembers["__require"].Ref)
 	r := c.renameSymbolsInChunk(chunk, chunkRepr.filesInChunkInOrder, timer)
@@ -4332,8 +4340,6 @@ func (c *linkerContext) generateChunkJS(chunks []chunkInfo, chunkIndex int, chun
 			partRange,
 			chunk.entryBits,
 			chunkAbsDir,
-			commonJSRef,
-			esmRef,
 			toModuleRef,
 			runtimeRequireRef,
 			compileResult,
@@ -4701,7 +4707,7 @@ func (c *linkerContext) generateGlobalNamePrefix() string {
 		join = ";"
 	}
 
-	if js_printer.CanQuoteIdentifier(prefix, c.options.UnsupportedJSFeatures, c.options.ASCIIOnly) {
+	if js_printer.CanEscapeIdentifier(prefix, c.options.UnsupportedJSFeatures, c.options.ASCIIOnly) {
 		if c.options.ASCIIOnly {
 			prefix = string(js_printer.QuoteIdentifier(nil, prefix, c.options.UnsupportedJSFeatures))
 		}
@@ -4713,7 +4719,7 @@ func (c *linkerContext) generateGlobalNamePrefix() string {
 
 	for _, name := range c.options.GlobalName[1:] {
 		oldPrefix := prefix
-		if js_printer.CanQuoteIdentifier(name, c.options.UnsupportedJSFeatures, c.options.ASCIIOnly) {
+		if js_printer.CanEscapeIdentifier(name, c.options.UnsupportedJSFeatures, c.options.ASCIIOnly) {
 			if c.options.ASCIIOnly {
 				name = string(js_printer.QuoteIdentifier(nil, name, c.options.UnsupportedJSFeatures))
 			}