+++ internal/js_parser/js_parser_lower.go
@@ -14,6 +14,19 @@ import (
 	"github.com/evanw/esbuild/internal/logger"
 )
 
+func (p *parser) prettyPrintTargetEnvironment(feature compat.JSFeature) (where string, notes []logger.MsgData) {
+	where = "the configured target environment"
+	if tsTarget := p.options.tsTarget; tsTarget != nil && tsTarget.UnsupportedJSFeatures.Has(feature) {
+		tracker := logger.MakeLineColumnTracker(&tsTarget.Source)
+		where = fmt.Sprintf("%s (%q)", where, tsTarget.Target)
+		notes = []logger.MsgData{logger.RangeData(&tracker, tsTarget.Range, fmt.Sprintf(
+			"The target environment was set to %q here", tsTarget.Target))}
+	} else if p.options.originalTargetEnv != "" {
+		where = fmt.Sprintf("%s (%s)", where, p.options.originalTargetEnv)
+	}
+	return
+}
+
 func (p *parser) markSyntaxFeature(feature compat.JSFeature, r logger.Range) (didGenerateError bool) {
 	didGenerateError = true
 
@@ -29,17 +42,7 @@ func (p *parser) markSyntaxFeature(feature compat.JSFeature, r logger.Range) (di
 	}
 
 	var name string
-	var notes []logger.MsgData
-	where := "the configured target environment"
-
-	if tsTarget := p.options.tsTarget; tsTarget != nil && tsTarget.UnsupportedJSFeatures.Has(feature) {
-		tracker := logger.MakeLineColumnTracker(&tsTarget.Source)
-		where = fmt.Sprintf("%s (%q)", where, tsTarget.Target)
-		notes = []logger.MsgData{logger.RangeData(&tracker, tsTarget.Range, fmt.Sprintf(
-			"The target environment was set to %q here", tsTarget.Target))}
-	} else if p.options.originalTargetEnv != "" {
-		where = fmt.Sprintf("%s (%s)", where, p.options.originalTargetEnv)
-	}
+	where, notes := p.prettyPrintTargetEnvironment(feature)
 
 	switch feature {
 	case compat.DefaultArgument:
@@ -1994,7 +1997,7 @@ func (p *parser) lowerClass(stmt js_ast.Stmt, expr js_ast.Expr, shadowRef js_ast
 				if kind == classKindExportDefaultStmt {
 					class.Name = &defaultName
 				} else {
-					class.Name = &js_ast.LocRef{Loc: classLoc, Ref: p.generateTempRef(tempRefNoDeclare, "zomzomz")}
+					class.Name = &js_ast.LocRef{Loc: classLoc, Ref: p.generateTempRef(tempRefNoDeclare, "")}
 				}
 			}
 			p.recordUsage(class.Name.Ref)
@@ -2568,25 +2571,17 @@ func (p *parser) lowerClass(stmt js_ast.Stmt, expr js_ast.Expr, shadowRef js_ast
 		p.recordUsage(nameForClassDecorators.Ref)
 	}
 	if generatedLocalStmt {
+		// "export default class x {}" => "class x {} export {x as default}"
 		if kind == classKindExportDefaultStmt {
-			// Generate a new default name symbol since the current one is being used
-			// by the class. If this SExportDefault turns into a variable declaration,
-			// we don't want it to accidentally use the same variable as the class and
-			// cause a name collision.
-			defaultRef := p.generateTempRef(tempRefNoDeclare, p.source.IdentifierName+"_default")
-			p.recordDeclaredSymbol(defaultRef)
-
-			name := nameFunc()
-			stmts = append(stmts, js_ast.Stmt{Loc: classLoc, Data: &js_ast.SExportDefault{
-				DefaultName: js_ast.LocRef{Loc: defaultName.Loc, Ref: defaultRef},
-				Value:       js_ast.Stmt{Loc: name.Loc, Data: &js_ast.SExpr{Value: name}},
+			stmts = append(stmts, js_ast.Stmt{Loc: classLoc, Data: &js_ast.SExportClause{
+				Items: []js_ast.ClauseItem{{Alias: "default", Name: defaultName}},
 			}})
 		}
 
 		// Calling "nameFunc" will set the class name, but we don't want it to have
 		// one. If the class name was necessary, we would have already split it off
-		// into a "const" symbol above. Reset it back to empty here now that we
-		// know we won't call "nameFunc" after this point.
+		// into a variable above. Reset it back to empty here now that we know we
+		// won't call "nameFunc" after this point.
 		class.Name = nil
 	}
 	return stmts, js_ast.Expr{}
@@ -2598,46 +2593,40 @@ func (p *parser) lowerTemplateLiteral(loc logger.Loc, e *js_ast.ETemplate) js_as
 		var value js_ast.Expr
 
 		// Handle the head
-		if len(e.HeadCooked) == 0 {
-			// "`${x}y`" => "x + 'y'"
-			part := e.Parts[0]
-			value = js_ast.Expr{Loc: loc, Data: &js_ast.EBinary{
-				Op:   js_ast.BinOpAdd,
-				Left: part.Value,
-				Right: js_ast.Expr{Loc: part.TailLoc, Data: &js_ast.EString{
-					Value:          part.TailCooked,
-					LegacyOctalLoc: e.LegacyOctalLoc,
-				}},
-			}}
-			e.Parts = e.Parts[1:]
-		} else {
-			// "`x${y}`" => "'x' + y"
-			value = js_ast.Expr{Loc: loc, Data: &js_ast.EString{
-				Value:          e.HeadCooked,
-				LegacyOctalLoc: e.LegacyOctalLoc,
-			}}
-		}
+		value = js_ast.Expr{Loc: loc, Data: &js_ast.EString{
+			Value:          e.HeadCooked,
+			LegacyOctalLoc: e.LegacyOctalLoc,
+		}}
 
-		// Handle the tail
+		// Handle the tail. Each one is handled with a separate call to ".concat()"
+		// to handle various corner cases in the specification including:
+		//
+		//   * For objects, "toString" must be called instead of "valueOf"
+		//   * Side effects must happen inline instead of at the end
+		//   * Passing a "Symbol" instance should throw
+		//
 		for _, part := range e.Parts {
-			value = js_ast.Expr{Loc: loc, Data: &js_ast.EBinary{
-				Op:    js_ast.BinOpAdd,
-				Left:  value,
-				Right: part.Value,
-			}}
+			var args []js_ast.Expr
 			if len(part.TailCooked) > 0 {
-				value = js_ast.Expr{Loc: loc, Data: &js_ast.EBinary{
-					Op:    js_ast.BinOpAdd,
-					Left:  value,
-					Right: js_ast.Expr{Loc: part.TailLoc, Data: &js_ast.EString{Value: part.TailCooked}},
-				}}
+				args = []js_ast.Expr{part.Value, {Loc: part.TailLoc, Data: &js_ast.EString{Value: part.TailCooked}}}
+			} else {
+				args = []js_ast.Expr{part.Value}
 			}
+			value = js_ast.Expr{Loc: loc, Data: &js_ast.ECall{
+				Target: js_ast.Expr{Loc: loc, Data: &js_ast.EDot{
+					Target:  value,
+					Name:    "concat",
+					NameLoc: part.Value.Loc,
+				}},
+				Args: args,
+			}}
 		}
 
 		return value
 	}
 
 	// Otherwise, call the tag with the template object
+	needsRaw := false
 	cooked := []js_ast.Expr{}
 	raw := []js_ast.Expr{}
 	args := make([]js_ast.Expr, 0, 1+len(e.Parts))
@@ -2646,8 +2635,12 @@ func (p *parser) lowerTemplateLiteral(loc logger.Loc, e *js_ast.ETemplate) js_as
 	// Handle the head
 	if e.HeadCooked == nil {
 		cooked = append(cooked, js_ast.Expr{Loc: e.HeadLoc, Data: js_ast.EUndefinedShared})
+		needsRaw = true
 	} else {
 		cooked = append(cooked, js_ast.Expr{Loc: e.HeadLoc, Data: &js_ast.EString{Value: e.HeadCooked}})
+		if !js_lexer.UTF16EqualsString(e.HeadCooked, e.HeadRaw) {
+			needsRaw = true
+		}
 	}
 	raw = append(raw, js_ast.Expr{Loc: e.HeadLoc, Data: &js_ast.EString{Value: js_lexer.StringToUTF16(e.HeadRaw)}})
 
@@ -2656,18 +2649,39 @@ func (p *parser) lowerTemplateLiteral(loc logger.Loc, e *js_ast.ETemplate) js_as
 		args = append(args, part.Value)
 		if part.TailCooked == nil {
 			cooked = append(cooked, js_ast.Expr{Loc: part.TailLoc, Data: js_ast.EUndefinedShared})
+			needsRaw = true
 		} else {
 			cooked = append(cooked, js_ast.Expr{Loc: part.TailLoc, Data: &js_ast.EString{Value: part.TailCooked}})
+			if !js_lexer.UTF16EqualsString(part.TailCooked, part.TailRaw) {
+				needsRaw = true
+			}
 		}
 		raw = append(raw, js_ast.Expr{Loc: part.TailLoc, Data: &js_ast.EString{Value: js_lexer.StringToUTF16(part.TailRaw)}})
 	}
 
 	// Construct the template object
-	args[0] = p.callRuntime(e.HeadLoc, "__template", []js_ast.Expr{
-		{Loc: e.HeadLoc, Data: &js_ast.EArray{Items: cooked, IsSingleLine: true}},
-		{Loc: e.HeadLoc, Data: &js_ast.EArray{Items: raw, IsSingleLine: true}},
-	})
+	cookedArray := js_ast.Expr{Loc: e.HeadLoc, Data: &js_ast.EArray{Items: cooked, IsSingleLine: true}}
+	var arrays []js_ast.Expr
+	if needsRaw {
+		arrays = []js_ast.Expr{cookedArray, {Loc: e.HeadLoc, Data: &js_ast.EArray{Items: raw, IsSingleLine: true}}}
+	} else {
+		arrays = []js_ast.Expr{cookedArray}
+	}
+	templateObj := p.callRuntime(e.HeadLoc, "__template", arrays)
 
+	// Cache it in a temporary object (required by the specification)
+	tempRef := p.generateTopLevelTempRef()
+	args[0] = js_ast.Expr{Loc: loc, Data: &js_ast.EBinary{
+		Op:   js_ast.BinOpLogicalOr,
+		Left: js_ast.Expr{Loc: loc, Data: &js_ast.EIdentifier{Ref: tempRef}},
+		Right: js_ast.Expr{Loc: loc, Data: &js_ast.EBinary{
+			Op:    js_ast.BinOpAssign,
+			Left:  js_ast.Expr{Loc: loc, Data: &js_ast.EIdentifier{Ref: tempRef}},
+			Right: templateObj,
+		}},
+	}}
+
+	// Call the tag function
 	return js_ast.Expr{Loc: loc, Data: &js_ast.ECall{
 		Target: e.TagOrNil,
 		Args:   args,